refactor: extract shared utilities from parsers
- Create utils/token-delta.ts with isAllZero() and toNonNegInt()
- Create utils/time.ts with coerceEpochMs()
- Update all parsers to use shared utilities
- Fix openclaw.ts to use toNonNegInt() for safer type coercion
- Update test imports for coerceEpochMs
