+++ packages/cli/src/parsers/opencode-session.ts
@@ -8,17 +8,7 @@
 import { readdir, readFile } from "node:fs/promises";
 import { join, basename } from "node:path";
 import type { SessionSnapshot, Source } from "@pew/core";
-
-/**
- * Coerce an epoch value to milliseconds.
- * Values < 1e12 are treated as seconds and multiplied by 1000.
- */
-function coerceEpochMs(v: unknown): number {
-  const n = Number(v);
-  if (!Number.isFinite(n) || n <= 0) return 0;
-  if (n < 1e12) return Math.floor(n * 1000);
-  return Math.floor(n);
-}
+import { coerceEpochMs } from "../utils/time.js";
 
 /**
  * Collect session snapshots from an OpenCode session directory.
