+++ packages/cli/src/parsers/opencode-sqlite-session.ts
@@ -8,6 +8,7 @@
 
 import type { SessionSnapshot, Source } from "@pew/core";
 import { hashProjectRef } from "../utils/hash-project-ref.js";
+import { coerceEpochMs } from "../utils/time.js";
 
 /** Row shape from the session table */
 export interface SessionRow {
@@ -24,17 +25,6 @@ export interface SessionMessageRow {
   role: string;
   time_created: number;
   data: string;
-}
-
-/**
- * Coerce an epoch value to milliseconds.
- * Values < 1e12 are treated as seconds and multiplied by 1000.
- */
-function coerceEpochMs(v: unknown): number {
-  const n = Number(v);
-  if (!Number.isFinite(n) || n <= 0) return 0;
-  if (n < 1e12) return Math.floor(n * 1000);
-  return Math.floor(n);
 }
 
 /**
